backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn==0.29.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# ML & Data
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4

# File handling
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.1